User: add optional $object_detail_level argument to getAll() and get(). Similar to Channel getAll() and get() this specifies the amount of detail that a server will return, see "Structure of a client object" at https://www.unrealircd.org/docs/JSON-RPC:User
Only works in UnrealIRCd 6.0.8 or later, simply ignored in earlier versions.

If you only need a minimal amount of information, consider using a
lower object_detail_level. It saves both CPU and bandwidth.

In UnrealIRCd 6.0.8+ the user.get call returns the same information with
the default object_detail_level as in 6.0.7.

For the user.list call, UnrealIRCd 6.0.8 has a "breaking change":
the "channels" of the user are no longer returned with the default
object_detail_level value of 2. If you require that information, then
set object_detail_level to 4, like: $rpc->user()->getAll(4);